issue104: merge planet url strategy classes.
