+++ app/src/main/java/org/apache/roller/weblogger/business/URLStrategy.java
@@ -14,6 +14,9 @@
  * limitations under the License.  For additional information regarding
  * copyright in this work, please see the NOTICE file in the top level
  * directory of this distribution.
+ *
+ * Source file modified from the original ASF source; all changes made
+ * are also under Apache License.
  */
 
 package org.apache.roller.weblogger.business;
@@ -245,4 +248,9 @@ public interface URLStrategy {
 
     /* Get URL for obtaining OAuth Access Token */
     String getOAuthAccessTokenURL();
+
+    /**
+     * Get URL of a planet's newsfeed.
+     */
+    String getPlanetURL(String planet);
 }
