+++ app/src/main/java/org/apache/roller/weblogger/business/Weblogger.java
@@ -14,6 +14,9 @@
  * limitations under the License.  For additional information regarding
  * copyright in this work, please see the NOTICE file in the top level
  * directory of this distribution.
+ *
+ * Source file modified from the original ASF source; all changes made
+ * are also under Apache License.
  */
 
 package org.apache.roller.weblogger.business;
@@ -181,5 +184,4 @@ public interface Weblogger {
 
 	PlanetManager getPlanetManager();
 
-	org.apache.roller.planet.business.PlanetURLStrategy getPlanetURLStrategy();
 }
